move formatting stuff over to GeneratePatchTargetHashes
