+++ tools/GeneratePatchTargetHashes/Program.cs
@@ -19,7 +19,7 @@ internal static class Program {
     var gameBinDir = new Uri(Path.Combine(gameRoot, "bin", "Win64_Shipping_Client")).LocalPath;
     Environment.CurrentDirectory = gameBinDir;
 
-    Assembly? AssemblyResolver(object sender, ResolveEventArgs eventArgs) {
+    Assembly AssemblyResolver(object sender, ResolveEventArgs eventArgs) {
       Console.WriteLine("Loading: " + eventArgs.Name);
 
       if (eventArgs.Name == null)
